- fix gtest compiler bug - fix epoll memory leak in windows - fix tc_encoder crash bug - fix tc_clientsocket bug - fix tc_timer memory leak - fix tc_openssl mode not be used - fix TC_ Http2 failure problem (header cannot use temporary variables) - always set thread data host info - Adding the initialization of enumeration value in the resetDefautlt function - Part of the file encoding format is adjusted to utf8 BOM to ensure that there is no problem in Windows compilation - tc_common add various date processing functions - tc_common add floating point comparison function - tc_thread add multi thread hash logic - tc_port add environment variable acquisition and system event registration callback (compatible with multi platform) - Adjust tc_sem_mutex event name (under Windows) to facilitate subsequent maintenance - rpc add support for http/sock4/sock5 proxy - application add ssl bind check - Fix the bug of RPC in HTTP short SSL package connection  (under Windows) - communicator add update endpoint - fix communicator exit logic to avoid possible crash in exit - optimize the memory copy problem, reduce the memory copy, and improve the performance - Adjust the custom callback and put it in servantproxy

+++ util/src/tc_bitmap.cpp
@@ -144,7 +144,7 @@ int TC_BitMap::BitMap::load5file(const string &sFile)
     char *pBuffer  = new char[iSize];
     while(true)
     {
-        int ret = fread(pBuffer, 1, iSize, fp); 
+        size_t ret = fread(pBuffer, 1, iSize, fp); 
         if(ret == 0)
         {
             break;
